닉네임 엔드포인트 컨트롤러 멤버 컨트롤러로 통합
