+++ src/main/kotlin/net/noti_me/dymit/dymit_backend_api/controllers/member/MemberController.kt
@@ -67,5 +67,10 @@ class MemberController(
         )
         return MemberCreateResponse.from(result)
     }
+
+    override fun checkNickname(nickname: String) {
+        logger.debug("checkNickname called with nickname: $nickname")
+        return memberCreateUsecase.checkNickname(nickname)
+    }
 }
